+++ src/0.signals/signals.c
@@ -2,29 +2,6 @@
 
 int	g_signal;
 
-void sigint_handler_heredoc(int signal)
-{
-	if (signal == SIGINT)
-	{
-		g_signal = signal;
-		close(STDIN_FILENO);
-	}
-}
-
-void	sigint_handler_main(int signal)
-{
-	if (signal == SIGINT)
-	{
-		write(STDOUT_FILENO, "\n", 1);
-		rl_on_new_line();
-		rl_replace_line("", 0);
-		rl_redisplay();
-	}
-	// g_signal = signal;
-}
-
-//signal handlers
-
 void	init_signals(struct sigaction *sa)
 {
 	g_signal = 0;
@@ -34,13 +11,13 @@ void	init_signals(struct sigaction *sa)
 	sigaction(SIGINT, sa, NULL);
 }
 
-void	ignore_signal_action(struct sigaction *sa)
+void	signal_action_ignore(struct sigaction *sa)
 {
 	sa->sa_handler = SIG_IGN;
 	sigaction(SIGINT, sa, NULL);
 }
 
-void	restart_signal_action_main(struct sigaction *sa)
+void	signal_action_main(struct sigaction *sa)
 {
 	g_signal = 0;
 	sa->sa_handler = sigint_handler_main;
@@ -49,10 +26,18 @@ void	restart_signal_action_main(struct sigaction *sa)
 	sigaction(SIGQUIT, sa, NULL);
 }
 
-void	restart_signal_action_heredoc(struct sigaction *sa)
+void	signal_action_heredoc(struct sigaction *sa)
 {
 	sa->sa_handler = sigint_handler_heredoc;
 	sigaction(SIGINT, sa, NULL);
 	sa->sa_handler = SIG_IGN;
+	sigaction(SIGQUIT, sa, NULL);
+}
+
+void	signal_action_child(struct sigaction *sa)
+{
+	sa->sa_handler = ;
+	sigaction(SIGINT, sa, NULL);
+	sa->sa_handler = ;
 	sigaction(SIGQUIT, sa, NULL);
 }
